104 add play all (#105)
* Includes a first pass for the in browser playlist playing functionality. Though I think it still has a memory leak somewhere.
Also added the playlist user service and the playlist controller.

* Added my basic components for managing playlists

* Playlists fetching to the app context. Updated the song screen to be able to assign songs add songs to playlists.

* Added service and endpoint to adjust the song ordering in a playlist.
Swapped out my base url and started using the vite proxy.
Refactored the playlist calls file to return a big object that has all the calls,
so that we have a consistent naming convention for the api calls.

* Rewrote the playlist listener to use the audio html elements again.
Fiddled with restoring the with keyword in the download_song endpoint